fix: adicionando regras para usuários inativos

+++ src/modules/User/services/deleteUser/DeleteUserService.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from "tsyringe";
+import { AppError } from "../../../../shared/error/AppError";
 import { IUserRepository } from "../../repositories/IUserRepository";
 
 @injectable()
@@ -10,6 +11,10 @@ class DeleteUserService {
 
     async execute(id: string): Promise<void> {
         /* Pendente aplicação das validações */
+        const wUserExists = await this._wRepository.findById(id);
+
+        if (!wUserExists || wUserExists.boInativo == 1) throw new AppError('Este usuário está inativo!');
+        
         this._wRepository.delete(id);
     }
 }
